c bon plus que les main

+++ ft_strlcpy.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int	ft_strlcpy(char *dest, const char *src, size_t size)
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
 	size_t	valeur;
